--- pages/genre.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const GenrePage = () => {
-  return (
-    <Page>
-      This is the genre page
-    </Page>
-  )
-}
-
-export default GenrePage;
-
-const Page = styled.main`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 80vh;
-`;
